moving shared components in atividade to its shared folder
