+++ app/atividades/shared/atividade-card/atividade-card.component.js
@@ -11,8 +11,8 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 const core_1 = require('@angular/core');
 const atividade_model_1 = require('../../shared/atividade.model');
 const material_1 = require('@angular/material');
-const visibility_confirmation_component_1 = require('./confirmation-dialogs/visibility-confirmation.component');
-const lock_confirmation_component_1 = require('./confirmation-dialogs/lock-confirmation.component');
+const visibility_confirmation_component_1 = require('../confirmation-dialogs/visibility-confirmation.component');
+const lock_confirmation_component_1 = require('../confirmation-dialogs/lock-confirmation.component');
 const material_2 = require('@angular/material');
 const atividade_service_1 = require('../../shared/atividade.service');
 let AtividadeCardComponent = class AtividadeCardComponent {
