+++ app/atividades/shared/atividade-card/atividade-card.component.scss
