+++ app/atividades/shared/atividade-card/snackbar-notification.component.js
