+++ app/atividades/shared/atividade-toolbar/atividade-toolbar.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
-import { SidenavService } from '../../core/sidenav/sidenav.service';
-import { MediaStateService } from '../../core/media-state.service';
+import { SidenavService } from '../../../core/sidenav/sidenav.service';
+import { MediaStateService } from '../../../core/media-state.service';
 import { Subscription } from 'rxjs/Subscription';
-import { SearchInputComponent } from '../../shared/search-input/search-input.component';
+import { SearchInputComponent } from '../../../shared/search-input/search-input.component';
 
 
 @Component ({
