+++ app/atividades/shared/confirmation-dialogs/delete-confirmation.component.js
