+++ app/atividades/shared/confirmation-dialogs/visibility-confirmation.component.js
