+++ app/atividades/shared/confirmation-dialogs/visibility-dialog.component.js
