+++ npm-debug.log
@@ -1,7 +1,7 @@
 0 info it worked if it ends with ok
 1 verbose cli [ '/usr/bin/nodejs', '/usr/bin/npm', 'start' ]
 2 info using npm@4.1.2
-3 info using node@v7.4.0
+3 info using node@v7.6.0
 4 verbose run-script [ 'prestart', 'start', 'poststart' ]
 5 info lifecycle angular-quickstart@1.0.0~prestart: angular-quickstart@1.0.0
 6 silly lifecycle angular-quickstart@1.0.0~prestart: no script for prestart, continuing
@@ -16,17 +16,17 @@
 14 verbose stack Exit status 2
 14 verbose stack     at EventEmitter.<anonymous> (/usr/lib/node_modules/npm/lib/utils/lifecycle.js:279:16)
 14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at EventEmitter.emit (events.js:192:7)
 14 verbose stack     at ChildProcess.<anonymous> (/usr/lib/node_modules/npm/lib/utils/spawn.js:40:14)
 14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:885:16)
+14 verbose stack     at ChildProcess.emit (events.js:192:7)
+14 verbose stack     at maybeClose (internal/child_process.js:890:16)
 14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
 15 verbose pkgid angular-quickstart@1.0.0
 16 verbose cwd /home/lucas/senec-app
 17 error Linux 4.5.0-040500-generic
 18 error argv "/usr/bin/nodejs" "/usr/bin/npm" "start"
-19 error node v7.4.0
+19 error node v7.6.0
 20 error npm  v4.1.2
 21 error code ELIFECYCLE
 22 error angular-quickstart@1.0.0 start: `tsc && concurrently "tsc -w" "lite-server" `
